+++ lcviz/tests/test_parser.py
@@ -81,7 +81,7 @@ def test_synthetic_lc(helper):
 def test_apply_xrangerois(helper, light_curve_like_kepler_quarter):
     lc = light_curve_like_kepler_quarter
     helper.load_data(lc)
-    viewer = helper.app.get_viewer(helper._default_time_viewer_reference_name)
+    viewer = helper.default_time_viewer._obj
     subset_plugin = helper.plugins['Subset Tools']
 
     # the min/max of temporal regions can be defined in two ways:
@@ -106,7 +106,7 @@ def test_apply_xrangerois(helper, light_curve_like_kepler_quarter):
 def test_apply_yrangerois(helper, light_curve_like_kepler_quarter):
     lc = light_curve_like_kepler_quarter
     helper.load_data(lc)
-    viewer = helper.app.get_viewer(helper._default_time_viewer_reference_name)
+    viewer = helper.default_time_viewer._obj
     subset_plugin = helper.plugins['Subset Tools']
 
     subset_plugin._obj.subset_selected = "Create New"