+++ lcviz/tests/test_plugin_markers.py
@@ -35,7 +35,7 @@ def test_docs_snippets(helper, light_curve_like_kepler_quarter):
 
 def test_plugin_markers(helper, light_curve_like_kepler_quarter):
     helper.load_data(light_curve_like_kepler_quarter)
-    tv = helper.app.get_viewer(helper._default_time_viewer_reference_name)
+    tv = helper.default_time_viewer._obj
 
     mp = helper.plugins['Markers']
     label_mouseover = mp._obj.coords_info