+++ CHANGES.rst
@@ -1,7 +1,12 @@
-0.2.0 - unreleased
+0.3.0 - unreleased
 ------------------
 
-* Clone viewer tool. [#74]
+* Ability to create additional viewers. [#94]
+
+0.2.0 (02-26-2024)
+------------------
+
+* Clone viewer tool. [#74, #91]
 
 * Flux column plugin to choose which column is treated as the flux column for each dataset. [#77]